+++ src/test/kotlin/io/snyk/plugin/cli/ConsoleCommandRunnerTest.kt
@@ -1,34 +1,25 @@
 package io.snyk.plugin.cli
 
+import com.intellij.credentialStore.Credentials
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.ide.plugins.PluginManagerCore
 import com.intellij.openapi.extensions.PluginId
-import com.intellij.openapi.progress.EmptyProgressIndicator
-import com.intellij.openapi.progress.ProgressIndicator
-import com.intellij.openapi.progress.ProgressManager
-import com.intellij.openapi.progress.Task
-import com.intellij.openapi.progress.impl.CoreProgressManager
-import com.intellij.openapi.util.registry.Registry
 import com.intellij.testFramework.LightPlatformTestCase
-import com.intellij.util.net.HttpConfigurable
-import io.mockk.every
-import io.mockk.mockkObject
+import com.intellij.util.net.ProxyConfiguration
+import com.intellij.util.net.ProxyConfiguration.StaticProxyConfiguration
+import com.intellij.util.net.ProxyCredentialStore
+import com.intellij.util.net.ProxySettings
 import io.mockk.unmockkAll
-import io.mockk.verify
-import io.sentry.protocol.SentryId
-import io.snyk.plugin.DEFAULT_TIMEOUT_FOR_SCAN_WAITING_MS
-import io.snyk.plugin.getPluginPath
 import io.snyk.plugin.pluginSettings
 import io.snyk.plugin.removeDummyCliFile
 import io.snyk.plugin.resetSettings
 import io.snyk.plugin.setupDummyCliFile
 import snyk.PLUGIN_ID
-import snyk.errorHandler.SentryErrorReporter
 import java.net.URLEncoder
-import java.util.Locale
 import java.util.UUID
-import java.util.concurrent.TimeUnit
 
+
+@Suppress("HttpUrlsUsage")
 class ConsoleCommandRunnerTest : LightPlatformTestCase() {
 
     override fun setUp() {
@@ -36,9 +27,6 @@ class ConsoleCommandRunnerTest : LightPlatformTestCase() {
         unmockkAll()
         resetSettings(project)
         setupDummyCliFile()
-        // don't report to Sentry when running this test
-        mockkObject(SentryErrorReporter)
-        every { SentryErrorReporter.captureException(any()) } returns SentryId()
     }
 
     override fun tearDown() {
@@ -150,17 +138,12 @@ class ConsoleCommandRunnerTest : LightPlatformTestCase() {
         }
     }
 
+    @Suppress("UnstableApiUsage")
     fun testSetupCliEnvironmentVariablesWithProxyWithoutAuth() {
-        val httpConfigurable = HttpConfigurable.getInstance()
-        val originalProxyHost = httpConfigurable.PROXY_HOST
-        val originalProxyPort = httpConfigurable.PROXY_PORT
-        val originalUseProxy = httpConfigurable.USE_HTTP_PROXY
-        val originalProxyExceptions = httpConfigurable.PROXY_EXCEPTIONS
+        val proxySettings = ProxySettings.getInstance()
+        val origConfig = proxySettings.getProxyConfiguration()
         try {
-            httpConfigurable.PROXY_PORT = 3128
-            httpConfigurable.PROXY_HOST = "testProxy"
-            httpConfigurable.USE_HTTP_PROXY = true
-            httpConfigurable.PROXY_EXCEPTIONS = "testExceptions"
+            proxySettings.setProxyConfiguration(createDummyProxyConfig())
 
             val generalCommandLine = GeneralCommandLine("")
             val expectedProxy = "http://testProxy:3128"
@@ -169,87 +152,38 @@ class ConsoleCommandRunnerTest : LightPlatformTestCase() {
 
             assertEquals(expectedProxy, generalCommandLine.environment["https_proxy"])
         } finally {
-            httpConfigurable.USE_HTTP_PROXY = originalUseProxy
-            httpConfigurable.PROXY_HOST = originalProxyHost
-            httpConfigurable.PROXY_PORT = originalProxyPort
-            httpConfigurable.PROXY_EXCEPTIONS = originalProxyExceptions
+            proxySettings.setProxyConfiguration(origConfig)
         }
     }
 
+    @Suppress("UnstableApiUsage")
     fun testSetupCliEnvironmentVariablesWithProxyWithAuth() {
-        val httpConfigurable = HttpConfigurable.getInstance()
-        val originalProxyHost = httpConfigurable.PROXY_HOST
-        val originalProxyPort = httpConfigurable.PROXY_PORT
-        val originalUseProxy = httpConfigurable.USE_HTTP_PROXY
-        val originalProxyAuthentication = httpConfigurable.PROXY_AUTHENTICATION
-        val originalLogin = httpConfigurable.proxyLogin
-        val originalPassword = httpConfigurable.plainProxyPassword
-        val originalProxyExceptions = httpConfigurable.PROXY_EXCEPTIONS
+        val proxySettings = ProxySettings.getInstance()
+        val credentialStore = ProxyCredentialStore.getInstance()
+        val origConfig = proxySettings.getProxyConfiguration()
+        val password = "test%!@Password"
+        val user = "testLogin"
         try {
-            httpConfigurable.PROXY_PORT = 3128
-            httpConfigurable.PROXY_HOST = "testProxy"
-            httpConfigurable.USE_HTTP_PROXY = true
-            httpConfigurable.PROXY_AUTHENTICATION = true
-            httpConfigurable.proxyLogin = "testLogin"
-            httpConfigurable.plainProxyPassword = "test%!@Password"
-            httpConfigurable.PROXY_EXCEPTIONS = "testExceptions"
-            val encodedLogin = URLEncoder.encode(httpConfigurable.proxyLogin, "UTF-8")
-            val encodedPassword = URLEncoder.encode(httpConfigurable.plainProxyPassword, "UTF-8")
+            val proxyConfiguration = createDummyProxyConfig()
+            proxySettings.setProxyConfiguration(proxyConfiguration)
+
+            val credentials = Credentials(user, password)
+            credentialStore.setCredentials(proxyConfiguration.host, proxyConfiguration.port, credentials, false)
+
+            val encodedLogin = URLEncoder.encode(user, "UTF-8")
+            val encodedPassword = URLEncoder.encode(password, "UTF-8")
 
             val generalCommandLine = GeneralCommandLine("")
             val expectedProxy =
                 "http://$encodedLogin:$encodedPassword@" +
-                    "${httpConfigurable.PROXY_HOST}:${httpConfigurable.PROXY_PORT}"
+                    "${proxyConfiguration.host}:${proxyConfiguration.port}"
 
             ConsoleCommandRunner().setupCliEnvironmentVariables(generalCommandLine, "")
 
             assertEquals(expectedProxy, generalCommandLine.environment["https_proxy"])
         } finally {
-            httpConfigurable.USE_HTTP_PROXY = originalUseProxy
-            httpConfigurable.PROXY_HOST = originalProxyHost
-            httpConfigurable.PROXY_PORT = originalProxyPort
-            httpConfigurable.PROXY_AUTHENTICATION = originalProxyAuthentication
-            httpConfigurable.proxyLogin = originalLogin
-            httpConfigurable.plainProxyPassword = originalPassword
-            httpConfigurable.PROXY_EXCEPTIONS = originalProxyExceptions
-        }
-    }
-
-    fun testSetupCliEnvironmentVariablesWithProxyWithCancelledAuth() {
-        val httpConfigurable = HttpConfigurable.getInstance()
-        val originalProxyHost = httpConfigurable.PROXY_HOST
-        val originalProxyPort = httpConfigurable.PROXY_PORT
-        val originalUseProxy = httpConfigurable.USE_HTTP_PROXY
-        val originalProxyAuthentication = httpConfigurable.PROXY_AUTHENTICATION
-        val originalAuthenticationCancelled = httpConfigurable.AUTHENTICATION_CANCELLED
-        val originalLogin = httpConfigurable.proxyLogin
-        val originalPassword = httpConfigurable.plainProxyPassword
-        val originalProxyExceptions = httpConfigurable.PROXY_EXCEPTIONS
-        try {
-            httpConfigurable.PROXY_PORT = 3128
-            httpConfigurable.PROXY_HOST = "testProxy"
-            httpConfigurable.USE_HTTP_PROXY = true
-            httpConfigurable.PROXY_AUTHENTICATION = true
-            httpConfigurable.AUTHENTICATION_CANCELLED = true
-            httpConfigurable.PROXY_EXCEPTIONS = "testExceptions"
-
-            val generalCommandLine = GeneralCommandLine("")
-            val expectedProxy =
-                "http://${httpConfigurable.PROXY_HOST}:${httpConfigurable.PROXY_PORT}"
-
-            ConsoleCommandRunner().setupCliEnvironmentVariables(generalCommandLine, "")
-
-            assertEquals(expectedProxy, generalCommandLine.environment["https_proxy"])
-            assertEquals(httpConfigurable.PROXY_EXCEPTIONS, generalCommandLine.environment["no_proxy"])
-        } finally {
-            httpConfigurable.USE_HTTP_PROXY = originalUseProxy
-            httpConfigurable.PROXY_HOST = originalProxyHost
-            httpConfigurable.PROXY_PORT = originalProxyPort
-            httpConfigurable.PROXY_AUTHENTICATION = originalProxyAuthentication
-            httpConfigurable.AUTHENTICATION_CANCELLED = originalAuthenticationCancelled
-            httpConfigurable.proxyLogin = originalLogin
-            httpConfigurable.plainProxyPassword = originalPassword
-            httpConfigurable.PROXY_EXCEPTIONS = originalProxyExceptions
+            proxySettings.setProxyConfiguration(origConfig)
+            credentialStore.clearAllCredentials()
         }
     }
 
@@ -266,37 +200,18 @@ class ConsoleCommandRunnerTest : LightPlatformTestCase() {
         assertEquals("2023.1".length, generalCommandLine.environment["SNYK_INTEGRATION_ENVIRONMENT_VERSION"]?.length)
     }
 
-    fun testErrorReportedWhenExecutionTimeoutExpire() {
-        val registryValue = Registry.get("snyk.timeout.results.waiting")
-        val defaultValue = registryValue.asInteger()
-        assertEquals(DEFAULT_TIMEOUT_FOR_SCAN_WAITING_MS, defaultValue)
-        registryValue.setValue(100)
-        val seconds = "3"
-        val sleepCommand = if (System.getProperty("os.name").lowercase(Locale.getDefault()).contains("win")) {
-            // see https://www.ibm.com/support/pages/timeout-command-run-batch-job-exits-immediately-and-returns-error-input-redirection-not-supported-exiting-process-immediately
-            listOf("ping", "-n", seconds, "localhost")
-        } else {
-            listOf("/bin/sleep", seconds)
+
+    private fun createDummyProxyConfig(): StaticProxyConfiguration {
+        val newConfig = object : StaticProxyConfiguration {
+            override val exceptions: String
+                get() = "testExceptions"
+            override val host: String
+                get() = "testProxy"
+            override val port: Int
+                get() = 3128
+            override val protocol: ProxyConfiguration.ProxyProtocol
+                get() = ProxyConfiguration.ProxyProtocol.HTTP
         }
-
-        val progressManager = ProgressManager.getInstance() as CoreProgressManager
-        val testRunFuture = progressManager.runProcessWithProgressAsynchronously(
-            object : Task.Backgroundable(project, "Test CLI command invocation", true) {
-                override fun run(indicator: ProgressIndicator) {
-                    val output = ConsoleCommandRunner().execute(sleepCommand, getPluginPath(), "", project)
-                    assertTrue(
-                        "Should get timeout error, but received:\n$output", output.startsWith("Execution timeout")
-                    )
-                }
-            },
-            EmptyProgressIndicator(),
-            null
-        )
-        testRunFuture.get(30000, TimeUnit.MILLISECONDS)
-
-        verify(exactly = 1) { SentryErrorReporter.captureException(any()) }
-
-        // clean up
-        registryValue.setValue(DEFAULT_TIMEOUT_FOR_SCAN_WAITING_MS)
+        return newConfig
     }
 }